refactor(bridge): extract callback and log registration
